Updates and Deletes for Leave Days

+++ database/migrations/2022_02_16_150613_add_dates_for_holidays_to_users_table.php
@@ -16,7 +16,7 @@ class AddDatesForHolidaysToUsersTable extends Migration
         Schema::table('users', function (Blueprint $table) {
             //
             $table->date('dateStarted')->default('2022-01-01')->after('name');
-            $table->integer('leaveDays')->default('20')->after('dateStarted');
+            $table->float('leaveDays')->default('20')->after('dateStarted');
 
         });
     }
